Move live_helper and error_helper to a single helper module.
