+++ installer/templates/phx_web/views/helpers.ex
@@ -1,9 +1,78 @@
-defmodule <%= @web_namespace %>.ErrorHelpers do
+defmodule <%= @web_namespace %>.Helpers do
   @moduledoc """
   Conveniences for translating and building error messages.
   """<%= if @html do %>
 
   use Phoenix.HTML
+
+  import Phoenix.LiveView
+  import Phoenix.LiveView.Helpers
+
+  alias Phoenix.LiveView.JS
+
+  @doc """
+  Renders a flash message.
+
+  The rendered flash receives a `:type` that will be used to define
+  proper classes to the element, and a `:message` which will be the
+  inner HTML, if any exists.
+
+  ## Examples
+
+      <.flash type="info" message="User created" />
+  """
+  def flash(assigns) do
+    if message = Map.get(assigns.flash, assigns.kind) do
+      ~H"""
+      <p class={"alert alert-#{@type}"} role="alert" phx-click="lv:clear-flash" phx-value-key={@kind}>
+       <%= @message %>
+      </p>
+      """
+    end
+  end
+
+  @doc """
+  Renders a live component inside a modal.
+
+  The rendered modal receives a `:return_to` option to properly update
+  the URL when the modal is closed.
+
+
+  """
+  def modal(assigns) do
+    assigns = assign_new(assigns, :return_to, fn -> nil end)
+
+    ~H"""
+    <div id="modal" class="phx-modal fade-in" phx-remove={hide_modal()}>
+      <div
+        id="modal-content"
+        class="phx-modal-content fade-in-scale"
+        phx-click-away={JS.dispatch("click", to: "#close")}
+        phx-window-keydown={JS.dispatch("click", to: "#close")}
+        phx-key="escape"
+      >
+        <%%= if @return_to do %>
+          <%%= live_patch "✖",
+            to: @return_to,
+            id: "close",
+            class: "phx-modal-close",
+            phx_click: hide_modal()
+          %>
+        <%% else %>
+         <a id="close" href="#" class="phx-modal-close" phx-click={hide_modal()}>✖</a>
+        <%% end %>
+
+        <%%= render_slot(@inner_block) %>
+      </div>
+    </div>
+    """
+  end
+
+  defp hide_modal(js \\ %JS{}) do
+    js
+    |> JS.hide(to: "#modal", transition: "fade-out")
+    |> JS.hide(to: "#modal-content", transition: "fade-out-scale")
+  end
 
   @doc """
   Generates tag for inlined form input errors.
