+++ priv/templates/phx.gen.json/changeset_view.ex
@@ -5,7 +5,7 @@ defmodule <%= inspect context.web_module %>.ChangesetView do
   Traverses and translates changeset errors.
 
   See `Ecto.Changeset.traverse_errors/2` and
-  `<%= inspect context.web_module %>.ErrorHelpers.translate_error/1` for more details.
+  `<%= inspect context.web_module %>.Helpers.translate_error/1` for more details.
   """
   def translate_errors(changeset) do
     Ecto.Changeset.traverse_errors(changeset, &translate_error/1)
